requirements_xlsx_converter.txt
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
odfpy==1.4.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
numpy==1.26.2 